requirements.txt
gensim==4.0.1
nltk==3.5
numpy==1.19.2
pandas==1.2.3
python-docx==0.8.11
scipy=1.6.0
scispacy=0.4.0
spacy==3.0.6
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_core_sci_sm-0.4.0.tar.gz